split of lvl_script_parser
