+++ src/creature_states_pray.h
@@ -45,6 +45,9 @@ short creature_sacrifice(struct Thing *thing);
 short creature_being_summoned(struct Thing *thing);
 
 void kill_all_players_chickens(PlayerNumber plyr_idx);
+
+TbBool find_temple_pool(int player_idx, struct Coord3d *pos);
+void process_sacrifice_creature(struct Coord3d *pos, int model, int owner, TbBool partial);
 /******************************************************************************/
 #ifdef __cplusplus
 }
