+++ src/lvl_script.c
@@ -79,6 +79,7 @@
 #include "lvl_script_lib.h"
 #include "lvl_script_statehandler.h"
 #include "lvl_script_conditions.h"
+#include "lvl_script_parser.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -88,234 +89,18 @@ extern "C" {
 
 extern void process_sacrifice_creature(struct Coord3d *pos, int model, int owner, TbBool partial);
 extern TbBool find_temple_pool(int player_idx, struct Coord3d *pos);
-extern void find_location_pos(long location, PlayerNumber plyr_idx, struct Coord3d *pos, const char *func_name);
+
 
 extern long near_map_block_creature_filter_diagonal_random(const struct Thing *thing, MaxTngFilterParam param, long maximizer);
 
 void script_process_value(unsigned long var_index, unsigned long plr_range_id, long val2, long val3, long val4, struct ScriptValue *value);
 
-
-
-const struct CommandDesc dk1_command_desc[];
-const struct CommandDesc subfunction_desc[] = {
-    {"RANDOM",                     "Aaaaaaaa", Cmd_RANDOM, NULL, NULL},
-    {"DRAWFROM",                   "Aaaaaaaa", Cmd_DRAWFROM, NULL, NULL},
-    {"IMPORT",                     "PA      ", Cmd_IMPORT, NULL, NULL},
-    {NULL,                         "        ", Cmd_NONE, NULL, NULL},
-  };
-
-
-const struct NamedCommand player_desc[] = {
-  {"PLAYER0",          PLAYER0},
-  {"PLAYER1",          PLAYER1},
-  {"PLAYER2",          PLAYER2},
-  {"PLAYER3",          PLAYER3},
-  {"PLAYER_GOOD",      PLAYER_GOOD},
-  {"ALL_PLAYERS",      ALL_PLAYERS},
-  {"PLAYER_NEUTRAL",   PLAYER_NEUTRAL},
-  {NULL,               0},
-};
+extern const struct CommandDesc command_desc[];
+extern const struct CommandDesc dk1_command_desc[];
 
 
 
-const struct NamedCommand controls_variable_desc[] = {
-    {"TOTAL_DIGGERS",               SVar_CONTROLS_TOTAL_DIGGERS},
-    {"TOTAL_CREATURES",             SVar_CONTROLS_TOTAL_CREATURES},
-    {"TOTAL_DOORS",                 SVar_TOTAL_DOORS},
-    {"TOTAL_AREA",                  SVar_TOTAL_AREA},
-    {"GOOD_CREATURES",              SVar_CONTROLS_GOOD_CREATURES},
-    {"EVIL_CREATURES",              SVar_CONTROLS_EVIL_CREATURES},
-    {NULL,                           0},
-};
 
-const struct NamedCommand comparison_desc[] = {
-  {"==",     MOp_EQUAL},
-  {"!=",     MOp_NOT_EQUAL},
-  {"<",      MOp_SMALLER},
-  {">",      MOp_GREATER},
-  {"<=",     MOp_SMALLER_EQ},
-  {">=",     MOp_GREATER_EQ},
-  {NULL,     0},
-};
-
-const struct NamedCommand head_for_desc[] = {
-  {"ACTION_POINT",         MLoc_ACTIONPOINT},
-  {"DUNGEON",              MLoc_PLAYERSDUNGEON},
-  {"DUNGEON_HEART",        MLoc_PLAYERSHEART},
-  {"APPROPIATE_DUNGEON",   MLoc_APPROPRTDUNGEON},
-  {NULL,                   0},
-};
-
-const struct NamedCommand timer_desc[] = {
-  {"TIMER0", 0},
-  {"TIMER1", 1},
-  {"TIMER2", 2},
-  {"TIMER3", 3},
-  {"TIMER4", 4},
-  {"TIMER5", 5},
-  {"TIMER6", 6},
-  {"TIMER7", 7},
-  {NULL,     0},
-};
-
-const struct NamedCommand flag_desc[] = {
-  {"FLAG0",  0},
-  {"FLAG1",  1},
-  {"FLAG2",  2},
-  {"FLAG3",  3},
-  {"FLAG4",  4},
-  {"FLAG5",  5},
-  {"FLAG6",  6},
-  {"FLAG7",  7},
-  {NULL,     0},
-};
-
-const struct NamedCommand hero_objective_desc[] = {
-  {"STEAL_GOLD",           CHeroTsk_StealGold},
-  {"STEAL_SPELLS",         CHeroTsk_StealSpells},
-  {"ATTACK_ENEMIES",       CHeroTsk_AttackEnemies},
-  {"ATTACK_DUNGEON_HEART", CHeroTsk_AttackDnHeart},
-  {"ATTACK_ROOMS",         CHeroTsk_AttackRooms},
-  {"DEFEND_PARTY",         CHeroTsk_DefendParty},
-  {"DEFEND_LOCATION",      CHeroTsk_DefendSpawn},
-  {"DEFEND_HEART",         CHeroTsk_DefendHeart},
-  {"DEFEND_ROOMS",         CHeroTsk_DefendRooms},
-  {NULL,                   0},
-};
-
-const struct NamedCommand msgtype_desc[] = {
-  {"SPEECH",           1},
-  {"SOUND",            2},
-  {NULL,               0},
-};
-
-const struct NamedCommand tendency_desc[] = {
-  {"IMPRISON",         1},
-  {"FLEE",             2},
-  {NULL,               0},
-};
-
-const struct NamedCommand creature_select_criteria_desc[] = {
-  {"MOST_EXPERIENCED",     CSelCrit_MostExperienced},
-  {"MOST_EXP_WANDERING",   CSelCrit_MostExpWandering},
-  {"MOST_EXP_WORKING",     CSelCrit_MostExpWorking},
-  {"MOST_EXP_FIGHTING",    CSelCrit_MostExpFighting},
-  {"LEAST_EXPERIENCED",    CSelCrit_LeastExperienced},
-  {"LEAST_EXP_WANDERING",  CSelCrit_LeastExpWandering},
-  {"LEAST_EXP_WORKING",    CSelCrit_LeastExpWorking},
-  {"LEAST_EXP_FIGHTING",   CSelCrit_LeastExpFighting},
-  {"NEAR_OWN_HEART",       CSelCrit_NearOwnHeart},
-  {"NEAR_ENEMY_HEART",     CSelCrit_NearEnemyHeart},
-  {"ON_ENEMY_GROUND",      CSelCrit_OnEnemyGround},
-  {"ON_FRIENDLY_GROUND",   CSelCrit_OnFriendlyGround},
-  {"ANYWHERE",             CSelCrit_Any},
-  {NULL,                   0},
-};
-
-const struct NamedCommand door_config_desc[] = {
-  {"ManufactureLevel",      1},
-  {"ManufactureRequired",   2},
-  {"Health",                3},
-  {"SellingValue",          4},
-  {"NametextId",            5},
-  {"TooltipTextId",         6},
-  {"Crate",                 7},
-  {"SymbolSprites",         8},
-  {"PointerSprites",        9},
-  {"PanelTabIndex",        10},
-  {NULL,                    0},
-};
-
-const struct NamedCommand trap_config_desc[] = {
-  {"NameTextID",           1},
-  {"TooltipTextID",        2},
-  {"SymbolSprites",        3},
-  {"PointerSprites",       4},
-  {"PanelTabIndex",        5},
-  {"Crate",                6},
-  {"ManufactureLevel",     7},
-  {"ManufactureRequired",  8},
-  {"Shots",                9},
-  {"TimeBetweenShots",    10},
-  {"SellingValue",        11},
-  {"Model",               12},
-  {"ModelSize",           13},
-  {"AnimationSpeed",      14},
-  {"TriggerType",         15},
-  {"ActivationType",      16},
-  {"EffectType",          17},
-  {"Hidden",              18},
-  {"TriggerAlarm",        19},
-  {"Slappable",           20},
-  {"Unanimated",          21},
-  {NULL,                   0},
-};
-
-/**
- * Text names of groups of GUI Buttons.
- */
-const struct NamedCommand gui_button_group_desc[] = {
-  {"MINIMAP",         GID_MINIMAP_AREA},
-  {"TABS",            GID_TABS_AREA},
-  {"INFO",            GID_INFO_PANE},
-  {"ROOM",            GID_ROOM_PANE},
-  {"POWER",           GID_POWER_PANE},
-  {"TRAP",            GID_TRAP_PANE},
-  {"DOOR",            GID_DOOR_PANE},
-  {"CREATURE",        GID_CREATR_PANE},
-  {"MESSAGE",         GID_MESSAGE_AREA},
-  {NULL,               0},
-};
-
-/**
- * Text names of campaign flags.
- */
-const struct NamedCommand campaign_flag_desc[] = {
-  {"CAMPAIGN_FLAG0",  0},
-  {"CAMPAIGN_FLAG1",  1},
-  {"CAMPAIGN_FLAG2",  2},
-  {"CAMPAIGN_FLAG3",  3},
-  {"CAMPAIGN_FLAG4",  4},
-  {"CAMPAIGN_FLAG5",  5},
-  {"CAMPAIGN_FLAG6",  6},
-  {"CAMPAIGN_FLAG7",  7},
-  {NULL,     0},
-};
-
-const struct NamedCommand script_operator_desc[] = {
-  {"SET",         1},
-  {"INCREASE",    2},
-  {"DECREASE",    3},
-  {"MULTIPLY",    4},
-  {NULL,          0},
-};
-
-#define DEALLOCATE_SCRIPT_VALUE \
-    if (value != &tmp_value) \
-    {                           \
-        value->flags = TrgF_DISABLED; \
-        gameadd.script.values_num--; \
-    }
-
-#define PROCESS_SCRIPT_VALUE(cmd) \
-    if ((get_script_current_condition() == CONDITION_ALWAYS) && (next_command_reusable == 0)) \
-    { \
-        script_process_value(cmd, 0, 0, 0, 0, value); \
-    }
-
-// For dynamic strings
-static char* script_strdup(const char *src)
-{
-    char *ret = gameadd.script.next_string;
-    int remain_len = sizeof(gameadd.script.strings) - (gameadd.script.next_string - gameadd.script.strings);
-    if (strlen(src) >= remain_len)
-    {
-        return NULL;
-    }
-    strcpy(ret, src);
-    gameadd.script.next_string += strlen(src) + 1;
-    return ret;
-}
 /******************************************************************************/
 static int filter_criteria_type(long desc_type)
 {
@@ -584,967 +369,7 @@ TbBool get_player_id_f(const char *plrname, long *plr_range_id, const char *func
     return true;
 }
 
-static void set_object_configuration_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    const char *objectname = scline->tp[0];
-    const char *property = scline->tp[1];
-    const char *new_value = scline->tp[2];
 
-    long objct_id = get_id(object_desc, objectname);
-    if (objct_id == -1)
-    {
-        SCRPTERRLOG("Unknown object, '%s'", objectname);
-        DEALLOCATE_SCRIPT_VALUE
-        return;
-    }
-
-    long number_value;
-    long objectvar = get_id(objects_object_commands, property);
-    if (objectvar == -1)
-    {
-        SCRPTERRLOG("Unknown object variable");
-        DEALLOCATE_SCRIPT_VALUE
-        return;
-    }
-    switch (objectvar)
-    {
-        case 2: // Genre
-            number_value = get_id(objects_genres_desc, new_value);
-            if (number_value == -1)
-            {
-                SCRPTERRLOG("Unknown object variable");
-                DEALLOCATE_SCRIPT_VALUE
-                return;
-            }
-            value->arg2 = number_value;
-            break;
-        case 3: // RELATEDCREATURE
-            number_value = get_id(creature_desc, new_value);
-            if (number_value == -1)
-            {
-                SCRPTERRLOG("Unknown object variable");
-                DEALLOCATE_SCRIPT_VALUE
-                    return;
-            }
-            value->arg2 = number_value;
-            break;
-        case  5: // AnimId
-        {
-            struct Objects obj_tmp;
-            number_value = get_anim_id(new_value, &obj_tmp);
-            if (number_value == 0)
-            {
-                SCRPTERRLOG("Invalid animation id");
-                DEALLOCATE_SCRIPT_VALUE
-                return;
-            }
-
-            value->str2 = script_strdup(new_value);
-            if (value->str2 == NULL)
-            {
-                SCRPTERRLOG("Run out script strings space");
-                DEALLOCATE_SCRIPT_VALUE
-                return;
-            }
-            break;
-        }
-        case 18: // MapIcon
-        {
-            number_value = get_icon_id(new_value);
-            if (number_value < 0)
-            {
-                SCRPTERRLOG("Invalid icon id");
-                DEALLOCATE_SCRIPT_VALUE
-                return;
-            }
-            value->arg2 = number_value;
-            break;
-        }
-        default:
-            value->arg2 = atoi(new_value);
-    }
-
-    SCRIPTDBG(7, "Setting object %s property %s to %d", objectname, property, number_value);
-    value->arg0 = objct_id;
-    value->arg1 = objectvar;
-
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void set_creature_configuration_check(const struct ScriptLine* scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-
-    short creatvar = get_id(creatmodel_attributes_commands, scline->tp[1]);
-    if (creatvar == -1)
-    {
-        SCRPTERRLOG("Unknown creature attribute");
-        DEALLOCATE_SCRIPT_VALUE
-            return;
-    }
-
-    if (creatvar == 20) // ATTACKPREFERENCE
-    {
-        value->shorts[2] = get_id(attackpref_desc, scline->tp[2]);
-    }
-    else
-    {
-        value->shorts[2] = atoi(scline->tp[2]);
-    }
-
-    SCRIPTDBG(7, "Setting creature %s attribute %d to %d (%d)", creature_code_name(scline->np[0]), scline->np[1], scline->np[2], scline->np[3]);
-    value->shorts[0] = scline->np[0];
-    value->shorts[1] = creatvar;
-    value->shorts[3] = scline->np[3];
-
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-
-static void set_creature_configuration_process(struct ScriptContext* context)
-{
-    short creatid = context->value->shorts[0];
-    struct CreatureStats* crstat = creature_stats_get(creatid);
-    struct CreatureModelConfig* crconf = &gameadd.crtr_conf.model[creatid];
-
-    short attribute = context->value->shorts[1];
-    short value = context->value->shorts[2];
-    short value2 = context->value->shorts[3];
-
-    switch (attribute)
-    {
-    case 1: // NAME
-        CONFWRNLOG("Attribute (%d) not supported", attribute);
-        break;
-    case 2: // HEALTH
-        crstat->health = value;
-        break;
-    case 3: // HEALREQUIREMENT
-        crstat->heal_requirement = value;
-        break;
-    case 4: // HEALTHRESHOLD
-        crstat->heal_threshold = value;
-        break;
-    case 5: // STRENGTH
-        crstat->strength = value;
-        break;
-    case 6: // ARMOUR
-        crstat->armour = value;
-        break;
-    case 7: // DEXTERITY
-        crstat->dexterity = value;
-        break;
-    case 8: // FEARWOUNDED
-        crstat->fear_wounded = value;
-        break;
-    case 9: // FEARSTRONGER
-        crstat->fear_stronger = value;
-        break;
-    case 10: // DEFENCE
-        crstat->defense = value;
-        break;
-    case 11: // LUCK
-        crstat->luck = value;
-        break;
-    case 12: // RECOVERY
-        crstat->sleep_recovery = value;
-        break;
-    case 13: // HUNGERRATE
-        crstat->hunger_rate = value;
-        break;
-    case 14: // HUNGERFILL
-        crstat->hunger_fill = value;
-        break;
-    case 15: // LAIRSIZE
-        crstat->lair_size = value;
-        break;
-    case 16: // HURTBYLAVA
-        crstat->hurt_by_lava = value;
-        break;
-    case 17: // BASESPEED
-        crstat->base_speed = value;
-        break;
-    case 18: // GOLDHOLD
-        crstat->gold_hold = value;
-        break;
-    case 19: // SIZE
-        crstat->size_xy = value;
-        crstat->size_yz = value2;
-        break;
-    case 20: // ATTACKPREFERENCE
-        //todo
-        crstat->attack_preference = value;
-        break;
-    case 21: // PAY
-        crstat->pay = value;
-        break;
-    case 22: // HEROVSKEEPERCOST
-        crstat->hero_vs_keeper_cost = value;
-        break;
-    case 23: // SLAPSTOKILL
-        crstat->slaps_to_kill = value;
-        break;
-    case 24: // CREATURELOYALTY
-    case 25: // LOYALTYLEVEL
-    case 28: // PROPERTIES
-        CONFWRNLOG("Attribute (%d) not supported", attribute);
-        break;
-    case 26: // DAMAGETOBOULDER
-        crstat->damage_to_boulder = value;
-        break;
-    case 27: // THINGSIZE
-        crstat->thing_size_xy = value;
-        crstat->thing_size_yz = value2;
-        break;
-    case 29: // NAMETEXTID
-        crconf->namestr_idx = value;
-        break;
-    case 30: // FEARSOMEFACTOR
-        crstat->fearsome_factor = value;
-        break;
-    case 31: // TOKINGRECOVERY
-        crstat->toking_recovery = value;
-        break;
-    case 0: // comment
-        break;
-    case -1: // end of buffer
-        break;
-    default:
-        CONFWRNLOG("Unrecognized command (%d)",attribute);
-        break;
-    }
-    check_and_auto_fix_stats();
-    creature_stats_updated(creatid);
-}
-
-static void set_object_configuration_process(struct ScriptContext *context)
-{
-    struct Objects* objdat = get_objects_data(context->value->arg0);
-    struct ObjectConfigStats* objst = &gameadd.object_conf.object_cfgstats[context->value->arg0];
-    struct ObjectConfig* objbc = &gameadd.object_conf.base_config[context->value->arg0];
-    switch (context->value->arg1)
-    {
-        case 2: // GENRE
-            objst->genre = context->value->arg2;
-            break;
-        case 3: // RELATEDCREATURE
-            objdat->related_creatr_model = context->value->arg2;
-            break;
-        case 4: // PROPERTIES
-            objst->model_flags = context->value->arg2;
-            break;
-        case 5: // ANIMATIONID
-            objdat->sprite_anim_idx = get_anim_id(context->value->str2, objdat);
-            break;
-        case 6: // ANIMATIONSPEED
-            objdat->anim_speed = context->value->arg2;
-            break;
-        case 7: //SIZE_XY
-            objdat->size_xy = context->value->arg2;
-            break;
-        case 8: // SIZE_YZ
-            objdat->size_yz = context->value->arg2;
-            break;
-        case 9: // MAXIMUMSIZE
-            objdat->sprite_size_max = context->value->arg2;
-            break;
-        case 10: // DESTROYONLIQUID
-            objdat->destroy_on_liquid = context->value->arg2;
-            break;
-        case 11: // DESTROYONLAVA
-            objdat->destroy_on_lava = context->value->arg2;
-            break;
-        case 12: // HEALTH
-            objbc->health = context->value->arg2;
-            break;
-        case 13: // FALLACCELERATION
-            objbc->fall_acceleration = context->value->arg2;
-            break;
-        case 14: // LIGHTUNAFFECTED
-            objbc->light_unaffected = context->value->arg2;
-            break;
-        case 15: // LIGHTINTENSITY
-            objbc->ilght.intensity = context->value->arg2;
-            break;
-        case 16: // LIGHTRADIUS
-            objbc->ilght.radius = context->value->arg2 << 8; //Mystery bit shift. Remove it to get divide by 0 errors.
-            break;
-        case 17: // LIGHTISDYNAMIC
-            objbc->ilght.is_dynamic = context->value->arg2;
-            break;
-        case 18: // MAPICON
-            objst->map_icon = context->value->arg2;
-            break;
-        default:
-            WARNMSG("Unsupported Object configuration, variable %d.", context->value->arg1);
-            break;
-    }
-    update_all_object_stats();
-}
-
-static void display_timer_check(const struct ScriptLine *scline)
-{
-    const char *timrname = scline->tp[1];
-    char timr_id = get_rid(timer_desc, timrname);
-    if (timr_id == -1)
-    {
-        SCRPTERRLOG("Unknown timer, '%s'", timrname);
-        return;
-    }
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->bytes[0] = (unsigned char)scline->np[0];
-    value->bytes[1] = timr_id;
-    value->arg1 = 0;
-    value->bytes[2] = (TbBool)scline->np[2];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-
-static void display_timer_process(struct ScriptContext *context)
-{
-    gameadd.script_player = context->value->bytes[0];
-    gameadd.script_timer_id = context->value->bytes[1];
-    gameadd.script_timer_limit = context->value->arg1;
-    gameadd.timer_real = context->value->bytes[2];
-    game.flags_gui |= GGUI_ScriptTimer;
-}
-
-static void add_to_timer_check(const struct ScriptLine *scline)
-{
-    const char *timrname = scline->tp[1];
-    long timr_id = get_rid(timer_desc, timrname);
-    if (timr_id == -1)
-    {
-        SCRPTERRLOG("Unknown timer, '%s'", timrname);
-        return;
-    }
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->arg0 = scline->np[0];
-    value->arg1 = timr_id;
-    value->arg2 = scline->np[2];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void add_to_timer_process(struct ScriptContext *context)
-{
-   add_to_script_timer(context->value->arg0, context->value->arg1, context->value->arg2);
-}
-
-static void add_bonus_time_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->arg0 = scline->np[0];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void add_bonus_time_process(struct ScriptContext *context)
-{
-   game.bonus_time += context->value->arg0;
-}
-
-static void display_variable_check(const struct ScriptLine *scline)
-{
-    long varib_id, varib_type;
-    if (!parse_get_varib(scline->tp[1], &varib_id, &varib_type))
-    {
-        SCRPTERRLOG("Unknown variable, '%s'", scline->tp[1]);
-        return;
-    }
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->bytes[0] = scline->np[0];
-    value->bytes[1] = scline->np[3];
-    gameadd.script_value_type = varib_type;
-    value->arg1 = varib_id;
-    value->arg2 = scline->np[2];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void display_variable_process(struct ScriptContext *context)
-{
-   gameadd.script_player = context->value->bytes[0];
-   gameadd.script_value_id = context->value->arg1;
-   gameadd.script_variable_target = context->value->arg2;
-   gameadd.script_variable_target_type = context->value->bytes[1];
-   game.flags_gui |= GGUI_Variable;
-}
-
-static void display_countdown_check(const struct ScriptLine *scline)
-{
-    if (scline->np[2] <= 0)
-    {
-        SCRPTERRLOG("Can't have a countdown to %ld turns.", scline->np[2]);
-        return;
-    }
-    const char *timrname = scline->tp[1];
-    char timr_id = get_rid(timer_desc, timrname);
-    if (timr_id == -1)
-    {
-        SCRPTERRLOG("Unknown timer, '%s'", timrname);
-        return;
-    }
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->bytes[0] = (unsigned char)scline->np[0];
-    value->bytes[1] = timr_id;
-    value->arg1 = scline->np[2];
-    value->bytes[2] = (TbBool)scline->np[3];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void cmd_no_param_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void hide_timer_process(struct ScriptContext *context)
-{
-   game.flags_gui &= ~GGUI_ScriptTimer;
-}
-
-static void hide_variable_process(struct ScriptContext *context)
-{
-   game.flags_gui &= ~GGUI_Variable;
-}
-
-static void create_effect_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    TbMapLocation location;
-    const char *effect_name = scline->tp[0];
-    long effct_id = get_rid(effect_desc, effect_name);
-    if (effct_id == -1)
-    {
-        if (parameter_is_number(effect_name))
-        {
-            effct_id = atoi(effect_name);
-        }
-        else
-        {
-            SCRPTERRLOG("Unrecognised effect: %s", effect_name);
-            return;
-        }
-    }
-    value->chars[0] = effct_id;
-    const char *locname = scline->tp[1];
-    if (!get_map_location_id(locname, &location))
-    {
-        return;
-    }
-    long stl_x;
-    long stl_y;
-    find_map_location_coords(location, &stl_x, &stl_y, 0, __func__);
-    value->bytes[1] = stl_x;
-    value->bytes[2] = stl_y;
-    value->arg1 = scline->np[2];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void create_effect_at_pos_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    const char *effect_name = scline->tp[0];
-    long effct_id = get_rid(effect_desc, effect_name);
-    if (effct_id == -1)
-    {
-        if (parameter_is_number(effect_name))
-        {
-            effct_id = atoi(effect_name);
-        }
-        else
-        {
-            SCRPTERRLOG("Unrecognised effect: %s", effect_name);
-            return;
-        }
-    }
-    value->chars[0] = effct_id;
-    if (subtile_coords_invalid(scline->np[1], scline->np[2]))
-    {
-        SCRPTERRLOG("Invalid co-ordinates: %ld, %ld", scline->np[1], scline->np[2]);
-        return;
-    }
-    value->bytes[1] = scline->np[1];
-    value->bytes[2] = scline->np[2];
-    value->arg1 = scline->np[3];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void create_effect_process(struct ScriptContext *context)
-{
-    struct Coord3d pos;
-    pos.x.stl.num = (MapSubtlCoord)context->value->bytes[1];
-    pos.y.stl.num = (MapSubtlCoord)context->value->bytes[2];
-    pos.z.val = get_floor_height(pos.x.stl.num, pos.y.stl.num);
-    TbBool Price = (context->value->chars[0] == -(TngEffElm_Price));
-    if (Price)
-    {
-        pos.z.val += 128;
-    }
-    struct Thing* efftng;
-    if (context->value->chars[0] >= 0)
-    {
-        efftng = create_effect(&pos, context->value->bytes[0], game.neutral_player_num);
-    }
-    else
-    {
-        efftng = create_effect_element(&pos, ~(context->value->bytes[0]) + 1, game.neutral_player_num);
-    }
-    if (!thing_is_invalid(efftng))
-    {
-        if (thing_in_wall_at(efftng, &efftng->mappos))
-        {
-            move_creature_to_nearest_valid_position(efftng);
-        }
-        if (Price)
-        {
-            efftng->long_13 = context->value->arg1;
-        }
-    }
-}
-
-static void set_heart_health_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->arg0 = scline->np[0];
-    if (scline->np[1] > (signed long)game.dungeon_heart_health)
-    {
-        SCRPTWRNLOG("Value %ld is greater than maximum: %ld", scline->np[1], game.dungeon_heart_health);
-        value->arg1 = game.dungeon_heart_health;
-    }
-    else
-    {
-        value->arg1 = scline->np[1];
-    }
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void set_heart_health_process(struct ScriptContext *context)
-{
-    struct Thing* heartng = get_player_soul_container(context->value->arg0);
-    if (!thing_is_invalid(heartng))
-    {
-        heartng->health = (short)context->value->arg1;
-    }
-}
-
-static void add_heart_health_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->arg0 = scline->np[0];
-    value->arg1 = scline->np[1];
-    value->arg2 = scline->np[2];
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void add_heart_health_process(struct ScriptContext *context)
-{
-    struct Thing* heartng = get_player_soul_container(context->value->arg0);
-    if (!thing_is_invalid(heartng))
-    {
-        short old_health = heartng->health;
-        long new_health = heartng->health + context->value->arg1;
-        if (new_health > (signed long)game.dungeon_heart_health)
-        {
-            SCRIPTDBG(7,"Player %d's calculated heart health (%ld) is greater than maximum: %ld", heartng->owner, new_health, game.dungeon_heart_health);
-            new_health = game.dungeon_heart_health;
-        }
-        heartng->health = (short)new_health;
-        TbBool warn_on_damage = (context->value->arg2);
-        if (warn_on_damage)
-        {
-            if (heartng->health < old_health)
-            {
-                event_create_event_or_update_nearby_existing_event(heartng->mappos.x.val, heartng->mappos.y.val, EvKind_HeartAttacked, heartng->owner, heartng->index);
-                if (is_my_player_number(heartng->owner))
-                {
-                    output_message(SMsg_HeartUnderAttack, 400, true);
-                }
-            }
-        }
-    }
-}
-
-static void heart_lost_quick_objective_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    if ((scline->np[0] < 0) || (scline->np[0] >= QUICK_MESSAGES_COUNT))
-    {
-        SCRPTERRLOG("Invalid QUICK OBJECTIVE number (%d)", scline->np[0]);
-        return;
-    }
-    if (strlen(scline->tp[1]) >= MESSAGE_TEXT_LEN)
-    {
-        SCRPTWRNLOG("Objective TEXT too long; truncating to %d characters", MESSAGE_TEXT_LEN-1);
-    }
-    if ((gameadd.quick_messages[scline->np[0]][0] != '\0') && (strcmp(gameadd.quick_messages[scline->np[0]],scline->tp[1]) != 0))
-    {
-        SCRPTWRNLOG("Quick Objective no %d overwritten by different text", scline->np[0]);
-    }
-    strncpy(gameadd.quick_messages[scline->np[0]], scline->tp[1], MESSAGE_TEXT_LEN-1);
-    gameadd.quick_messages[scline->np[0]][MESSAGE_TEXT_LEN-1] = '\0';
-    
-    TbMapLocation location;
-    if (scline->tp[2][0] != '\0')
-    {
-        get_map_location_id(scline->tp[2], &location);
-    }
-
-    value->arg0 = scline->np[0];
-    value->uarg2 = location;
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void heart_lost_quick_objective_process(struct ScriptContext *context)
-{
-    gameadd.heart_lost_display_message = true;
-    gameadd.heart_lost_quick_message = true;
-    gameadd.heart_lost_message_id = context->value->arg0;
-    gameadd.heart_lost_message_target = context->value->arg2;
-}
-
-static void heart_lost_objective_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-    value->arg0 = scline->np[0];
-    TbMapLocation location;
-    if (scline->tp[1][0] != '\0')
-    {
-        get_map_location_id(scline->tp[1], &location);
-    }
-    value->uarg1 = location;
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void heart_lost_objective_process(struct ScriptContext *context)
-{
-    gameadd.heart_lost_display_message = true;
-    gameadd.heart_lost_quick_message = false;
-    gameadd.heart_lost_message_id = context->value->arg0;
-    gameadd.heart_lost_message_target = context->value->arg1;
-}
-
-static void null_process(struct ScriptContext *context)
-{
-}
-
-
-static void delete_from_party_check(const struct ScriptLine *scline)
-{
-    int party_id = get_party_index_of_name(scline->tp[0]);
-    if (party_id < 0)
-    {
-        SCRPTERRLOG("Invalid Party:%s",scline->tp[0]);
-        return;
-    }
-    long creature_id = get_rid(creature_desc, scline->tp[1]);
-    if (creature_id == -1)
-    {
-      SCRPTERRLOG("Unknown creature, '%s'", scline->tp[1]);
-      return;
-    }
-    if ((get_script_current_condition() == CONDITION_ALWAYS) && (next_command_reusable == 0))
-    {
-        delete_member_from_party(party_id, creature_id, scline->np[2]);
-    } else
-    {
-        struct PartyTrigger* pr_trig = &gameadd.script.party_triggers[gameadd.script.party_triggers_num % PARTY_TRIGGERS_COUNT];
-        pr_trig->flags = TrgF_DELETE_FROM_PARTY;
-        pr_trig->flags |= next_command_reusable?TrgF_REUSABLE:0;
-        pr_trig->party_id = party_id;
-        pr_trig->creatr_id = creature_id;
-        pr_trig->crtr_level = scline->np[2];
-        pr_trig->condit_idx = get_script_current_condition();
-
-        gameadd.script.party_triggers_num++;
-    }
-}
-
-static void add_to_party_check(const struct ScriptLine *scline)
-{
-    int party_id = get_party_index_of_name(scline->tp[0]);
-    if (party_id < 0)
-    {
-        SCRPTERRLOG("Invalid Party:%s",scline->tp[1]);
-        return;
-    }
-    if ((scline->np[2] < 1) || (scline->np[2] > CREATURE_MAX_LEVEL))
-    {
-      SCRPTERRLOG("Invalid Creature Level parameter; %ld not in range (%d,%d)",scline->np[2],1,CREATURE_MAX_LEVEL);
-      return;
-    }
-    long crtr_id = get_rid(creature_desc, scline->tp[1]);
-    if (crtr_id == -1)
-    {
-      SCRPTERRLOG("Unknown creature, '%s'", scline->tp[1]);
-      return;
-    }
-    long objective_id = get_rid(hero_objective_desc, scline->tp[4]);
-    if (objective_id == -1)
-    {
-      SCRPTERRLOG("Unknown party member objective, '%s'", scline->tp[4]);
-      return;
-    }
-  //SCRPTLOG("Party '%s' member kind %d, level %d",prtname,crtr_id,crtr_level);
-
-    if ((get_script_current_condition() == CONDITION_ALWAYS) && (next_command_reusable == 0))
-    {
-        add_member_to_party(party_id, crtr_id, scline->np[2], scline->np[3], objective_id, scline->np[5]);
-    } else
-    {
-        struct PartyTrigger* pr_trig = &gameadd.script.party_triggers[gameadd.script.party_triggers_num % PARTY_TRIGGERS_COUNT];
-        pr_trig->flags = TrgF_ADD_TO_PARTY;
-        pr_trig->flags |= next_command_reusable?TrgF_REUSABLE:0;
-        pr_trig->party_id = party_id;
-        pr_trig->creatr_id = crtr_id;
-        pr_trig->crtr_level = scline->np[2];
-        pr_trig->carried_gold = scline->np[3];
-        pr_trig->objectv = objective_id;
-        pr_trig->countdown = scline->np[5];
-        pr_trig->condit_idx = get_script_current_condition();
-
-        gameadd.script.party_triggers_num++;
-    }
-}
-
-static void add_to_party_process(struct ScriptContext *context)
-{
-    struct PartyTrigger* pr_trig = context->pr_trig;
-    add_member_to_party(pr_trig->party_id, pr_trig->creatr_id, pr_trig->crtr_level, pr_trig->carried_gold, pr_trig->objectv, pr_trig->countdown);
-}
-
-static int sac_compare_fn(const void *ptr_a, const void *ptr_b)
-{
-    const char *a = (const char*)ptr_a;
-    const char *b = (const char*)ptr_b;
-    return *a < *b;
-}
-static void set_sacrifice_recipe_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-
-    value->sac.action = get_rid(rules_sacrifices_commands, scline->tp[0]);
-    if (value->sac.action == -1)
-    {
-        SCRPTERRLOG("Unexpected action:%s", scline->tp[0]);
-        return;
-    }
-    long param;
-    if ((value->sac.action == SacA_CustomPunish) || (value->sac.action == SacA_CustomReward))
-    {
-        param = get_id(flag_desc, scline->tp[1]) + 1;
-    }
-    else
-    {
-        param = get_id(creature_desc, scline->tp[1]);
-        if (param == -1)
-        {
-            param = get_id(sacrifice_unique_desc, scline->tp[1]);
-        }
-        if (param == -1)
-        {
-            param = get_id(spell_desc, scline->tp[1]);
-        }
-    }
-    if (param == -1 && (strcmp(scline->tp[1], "NONE") == 0))
-    {
-        param = 0;
-    }
-
-    if (param < 0)
-    {
-        param = 0;
-        value->sac.action = SacA_None;
-        SCRPTERRLOG("Unexpected parameter:%s", scline->tp[1]);
-    }
-    value->sac.param = param;
-
-    for (int i = 0; i < MAX_SACRIFICE_VICTIMS; i++)
-    {
-       long vi = get_rid(creature_desc, scline->tp[i + 2]);
-       if (vi < 0)
-         vi = 0;
-       value->sac.victims[i] = vi;
-    }
-    qsort(value->sac.victims, MAX_SACRIFICE_VICTIMS, sizeof(value->sac.victims[0]), &sac_compare_fn);
-
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void remove_sacrifice_recipe_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-
-    value->sac.action = SacA_None;
-    value->sac.param = 0;
-
-    for (int i = 0; i < MAX_SACRIFICE_VICTIMS; i++)
-    {
-       long vi = get_rid(creature_desc, scline->tp[i]);
-       if (vi < 0)
-         vi = 0;
-       value->sac.victims[i] = vi;
-    }
-    qsort(value->sac.victims, MAX_SACRIFICE_VICTIMS, sizeof(value->sac.victims[0]), &sac_compare_fn);
-
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void set_sacrifice_recipe_process(struct ScriptContext *context)
-{
-    long victims[MAX_SACRIFICE_VICTIMS];
-    struct Coord3d pos;
-    int action = context->value->sac.action;
-    int param = context->value->sac.param;
-    for (int i = 0; i < MAX_SACRIFICE_VICTIMS; i++)
-    {
-        victims[i] = context->value->sac.victims[i];
-    }
-    for (int i = 1; i < MAX_SACRIFICE_RECIPES; i++)
-    {
-        struct SacrificeRecipe* sac = &gameadd.sacrifice_recipes[i];
-        if (sac->action == (long)SacA_None)
-        {
-            break;
-        }
-        if (memcmp(victims, sac->victims, sizeof(victims)) == 0)
-        {
-            sac->action = action;
-            sac->param = param;
-            if (action == (long)SacA_None)
-            {
-                // remove empty space
-                memmove(sac, sac + 1, (MAX_SACRIFICE_RECIPES - 1 - (sac - &gameadd.sacrifice_recipes[0])) * sizeof(*sac));
-            }
-            return;
-        }
-    }
-    if (action == (long)SacA_None) // No rule found
-    {
-        WARNLOG("Unable to find sacrifice rule to remove");
-        return;
-    }
-    struct SacrificeRecipe* sac = get_unused_sacrifice_recipe_slot();
-    if (sac == &gameadd.sacrifice_recipes[0])
-    {
-        ERRORLOG("No free sacrifice rules");
-        return;
-    }
-    memcpy(sac->victims, victims, sizeof(victims));
-    sac->action = action;
-    sac->param = param;
-
-    if (find_temple_pool(context->player_idx, &pos))
-    {
-        // Check if sacrifice pool already matches
-        for (int i = 0; i < sizeof(victims); i++)
-        {
-            if (victims[i] == 0)
-                break;
-            process_sacrifice_creature(&pos, victims[i], context->player_idx, false);
-        }
-    }
-}
-
-static void set_box_tooltip(const struct ScriptLine *scline)
-{
-  if ((scline->np[0] < 0) || (scline->np[0] >= CUSTOM_BOX_COUNT))
-  {
-    SCRPTERRLOG("Invalid CUSTOM_BOX number (%ld)", scline->np[0]);
-    return;
-  }
-  int idx = scline->np[0];
-  if (strlen(scline->tp[1]) >= MESSAGE_TEXT_LEN)
-  {
-      SCRPTWRNLOG("Tooltip TEXT too long; truncating to %d characters", MESSAGE_TEXT_LEN-1);
-  }
-  if ((gameadd.box_tooltip[idx][0] != '\0') && (strcmp(gameadd.box_tooltip[idx], scline->tp[1]) != 0))
-  {
-      SCRPTWRNLOG("Box tooltip #%d overwritten by different text", idx);
-  }
-  strncpy(gameadd.box_tooltip[idx], scline->tp[1], MESSAGE_TEXT_LEN-1);
-  gameadd.box_tooltip[idx][MESSAGE_TEXT_LEN-1] = '\0';
-}
-
-static void set_box_tooltip_id(const struct ScriptLine *scline)
-{
-  if ((scline->np[0] < 0) || (scline->np[0] >= CUSTOM_BOX_COUNT))
-  {
-    SCRPTERRLOG("Invalid CUSTOM_BOX number (%ld)", scline->np[0]);
-    return;
-  }
-  int idx = scline->np[0];
-  strncpy(gameadd.box_tooltip[idx], get_string(scline->np[1]), MESSAGE_TEXT_LEN-1);
-  gameadd.box_tooltip[idx][MESSAGE_TEXT_LEN-1] = '\0';
-}
-
-// 1/4 turn minimal
-#define FX_LINE_TIME_PARTS 4
-
-static void create_effects_line_check(const struct ScriptLine *scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-
-    ((long*)(&value->bytes[0]))[0] = scline->np[0]; // AP `from`
-    ((long*)(&value->bytes[4]))[0] = scline->np[1]; // AP `to`
-    value->chars[8] = scline->np[2]; // curvature
-    value->bytes[9] = scline->np[3]; // spatial stepping
-    value->bytes[10] = scline->np[4]; // temporal stepping
-    // TODO: use effect elements when below zero?
-    value->chars[11] = scline->np[5]; // effect
-
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-static void create_effects_line_process(struct ScriptContext *context)
-{
-    struct ScriptFxLine *fx_line = NULL;
-    for (int i = 0; i < (sizeof(gameadd.fx_lines) / sizeof(gameadd.fx_lines[0])); i++)
-    {
-        if (!gameadd.fx_lines[i].used)
-        {
-            fx_line = &gameadd.fx_lines[i];
-            fx_line->used = true;
-            gameadd.active_fx_lines++;
-            break;
-        }
-    }
-    if (fx_line == NULL)
-    {
-        ERRORLOG("Too many fx_lines");
-        return;
-    }
-    find_location_pos(((long *)(&context->value->bytes[0]))[0], context->player_idx, &fx_line->from, __func__);
-    find_location_pos(((long *)(&context->value->bytes[4]))[0], context->player_idx, &fx_line->to, __func__);
-    fx_line->curvature = context->value->chars[8];
-    fx_line->spatial_step = context->value->bytes[9] * 32;
-    fx_line->steps_per_turn = context->value->bytes[10];
-    fx_line->effect = context->value->chars[11];
-    fx_line->here = fx_line->from;
-    fx_line->step = 0;
-
-    if (fx_line->steps_per_turn <= 0)
-    {
-        fx_line->steps_per_turn = 32 * 255; // whole map
-    }
-
-    int dx = fx_line->to.x.val - fx_line->from.x.val;
-    int dy = fx_line->to.y.val - fx_line->from.y.val;
-    if ((dx * dx + dy * dy) != 0)
-    {
-        float len = sqrt((float)dx * dx + dy * dy);
-        fx_line->total_steps = (int)(len / fx_line->spatial_step) + 1;
-
-        int d_cx = -dy * fx_line->curvature / 32;
-        int d_cy = +dx * fx_line->curvature / 32;
-        fx_line->cx = (fx_line->to.x.val + fx_line->from.x.val - d_cx)/2;
-        fx_line->cy = (fx_line->to.y.val + fx_line->from.y.val - d_cy)/2;
-    }
-    else
-    {
-      fx_line->total_steps = 1;
-    }
-    fx_line->partial_steps = FX_LINE_TIME_PARTS;
-}
 
 static void process_fx_line(struct ScriptFxLine *fx_line)
 {
@@ -1582,8 +407,6 @@ static void process_fx_line(struct ScriptFxLine *fx_line)
     }
 }
 
-
-
 static void player_reveal_map_area(PlayerNumber plyr_idx, long x, long y, long w, long h)
 {
   SYNCDBG(0,"Revealing around (%d,%d)",x,y);
@@ -1604,482 +427,10 @@ void player_reveal_map_location(int plyr_idx, TbMapLocation target, long r)
   reveal_map_area(plyr_idx, x-(r>>1), x+(r>>1)+(r&1), y-(r>>1), y+(r>>1)+(r&1));
 }
 
-static void display_objective_check(const struct ScriptLine *scline)
-{
-  long msg_num = scline->np[0];
-  long x, y;
-  TbMapLocation location = 0;
-  if ((msg_num < 0) || (msg_num >= STRINGS_MAX))
-  {
-    SCRPTERRLOG("Invalid TEXT number");
-    return;
-  }
-  if (scline->command == Cmd_DISPLAY_OBJECTIVE)
-  {
-    const char *where = scline->tp[1];
-    if (!get_map_location_id(where, &location))
-    {
-      return;
-    }
-    command_add_value(Cmd_DISPLAY_OBJECTIVE, ALL_PLAYERS, msg_num, location, 0);
-  }
-  else
-  {
-    x = scline->np[1];
-    y = scline->np[2];
-    command_add_value(Cmd_DISPLAY_OBJECTIVE, ALL_PLAYERS, msg_num, location, get_subtile_number(x,y));
-  }
-}
-
-static void display_objective_process(struct ScriptContext *context)
-{
-      if ( (my_player_number >= context->plr_start) && (my_player_number < context->plr_end) )
-      {
-          set_general_objective(context->value->arg0,
-              context->value->arg1,
-              stl_num_decode_x(context->value->arg2),
-              stl_num_decode_y(context->value->arg2));
-      }
-}
-
-static void conceal_map_rect_check(const struct ScriptLine *scline)
-{
-    TbBool all = strcmp(scline->tp[5], "ALL") == 0;
-    if (!all)
-        all = strcmp(scline->tp[5], "1") == 0;
-
-    command_add_value(Cmd_CONCEAL_MAP_RECT, scline->np[0], scline->np[1], scline->np[2],
-                      (scline->np[4]<<16) | scline->np[3] | (all?1<<24:0));
-}
-
-static void conceal_map_rect_process(struct ScriptContext *context)
-{
-    long w = context->value->bytes[8];
-    long h = context->value->bytes[10];
-
-    conceal_map_area(context->value->plyr_range, context->value->arg0 - (w>>1), context->value->arg0 + (w>>1) + (w&1),
-                     context->value->arg1 - (h>>1), context->value->arg1 + (h>>1) + (h&1), context->value->bytes[11]);
-}
-
-void refresh_trap_anim(long trap_id)
-{
-    int k = 0;
-    const struct StructureList* slist = get_list_for_thing_class(TCls_Trap);
-    int i = slist->index;
-    while (i != 0)
-    {
-        struct Thing* traptng = thing_get(i);
-        if (thing_is_invalid(traptng))
-        {
-            ERRORLOG("Jump to invalid thing detected");
-            break;
-        }
-        i = traptng->next_of_class;
-        // Per thing code
-        if (traptng->model == trap_id)
-        {
-            traptng->anim_sprite = gameadd.trap_stats[trap_id].sprite_anim_idx;
-            struct TrapStats* trapstat = &gameadd.trap_stats[traptng->model];
-            char start_frame;
-            if (trapstat->field_13) {
-                start_frame = -1;
-            }
-            else {
-                start_frame = 0;
-            }
-            set_thing_draw(traptng, trapstat->sprite_anim_idx, trapstat->anim_speed, trapstat->sprite_size_max, trapstat->unanimated, start_frame, 2);
-        }
-        // Per thing code ends
-        k++;
-        if (k > slist->index)
-        {
-            ERRORLOG("Infinite loop detected when sweeping things list");
-            break;
-        }
-    }
-}
-
-static void set_trap_configuration_check(const struct ScriptLine* scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-
-    const char *trapname = scline->tp[0];
-    short trap_id = get_id(trap_desc, trapname);
-    if (trap_id == -1)
-    {
-        SCRPTERRLOG("Unknown trap, '%s'", trapname);
-        DEALLOCATE_SCRIPT_VALUE
-        return;
-    }
-
-    short trapvar = get_id(trap_config_desc, scline->tp[1]);
-    if (trapvar == -1)
-    {
-        SCRPTERRLOG("Unknown trap variable");
-        DEALLOCATE_SCRIPT_VALUE
-        return;
-    }
-
-    value->shorts[0] = trap_id;
-    value->shorts[1] = trapvar;
-    if (trapvar == 3) // SymbolSprites
-    {
-        char *tmp = malloc(strlen(scline->tp[2]) + strlen(scline->tp[3]) + 3);
-        // Pass two vars along as one merged val like: first\nsecond\m
-        strcpy(tmp, scline->tp[2]);
-        strcat(tmp, "|");
-        strcat(tmp,scline->tp[3]);
-        value->str2 = script_strdup(tmp); // first\0second
-        value->str2[strlen(scline->tp[2])] = 0;
-        free(tmp);
-        if (value->str2 == NULL)
-        {
-            SCRPTERRLOG("Run out script strings space");
-            DEALLOCATE_SCRIPT_VALUE
-            return;
-        }
-    }
-    else if (
-        (trapvar != 4) && // PointerSprites
-        (trapvar != 12) // Model
-        )
-    {
-        if ((scline->np[2] > 0xFFFF) || (scline->np[2] < 0))
-        {
-            SCRPTERRLOG("Value out of range: %d", scline->np[2]);
-            DEALLOCATE_SCRIPT_VALUE
-            return;
-        }
-        value->shorts[2] = (short)scline->np[2];
-    }
-    else
-    {
-        value->str2 = script_strdup(scline->tp[2]);
-        if (value->str2 == NULL)
-        {
-            SCRPTERRLOG("Run out script strings space");
-            DEALLOCATE_SCRIPT_VALUE
-            return;
-        }
-    }
-    SCRIPTDBG(7, "Setting trap %s property %d to %d", trapname, trapvar, value->shorts[2]);
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-void set_trap_configuration_process(struct ScriptContext *context)
-{
-    long trap_type = context->value->shorts[0];
-    struct TrapConfigStats *trapst = &gameadd.trapdoor_conf.trap_cfgstats[trap_type];
-    struct ManfctrConfig *mconf = &gameadd.traps_config[trap_type];
-    struct ManufactureData *manufctr = get_manufacture_data(trap_type);
-    short value = context->value->shorts[2];
-    switch (context->value->shorts[1])
-    {
-        case 1: // NameTextID
-            trapst->name_stridx = value;
-            break;
-        case 2: // TooltipTextID
-            trapst->tooltip_stridx = value;
-            break;
-        case 3: // SymbolSprites
-        {
-            trapst->bigsym_sprite_idx = get_icon_id(context->value->str2); // First
-            trapst->medsym_sprite_idx = get_icon_id(context->value->str2 + strlen(context->value->str2) + 1); // Second
-            if (trapst->bigsym_sprite_idx < 0)
-                trapst->bigsym_sprite_idx = bad_icon_id;
-            if (trapst->medsym_sprite_idx < 0)
-                trapst->medsym_sprite_idx = bad_icon_id;
-            manufctr->bigsym_sprite_idx = trapst->bigsym_sprite_idx;
-            manufctr->medsym_sprite_idx = trapst->medsym_sprite_idx;
-            update_trap_tab_to_config();
-        }
-            break;
-        case 4: // PointerSprites
-            trapst->pointer_sprite_idx = get_icon_id(context->value->str2);
-            if (trapst->pointer_sprite_idx < 0)
-                trapst->pointer_sprite_idx = bad_icon_id;
-            update_trap_tab_to_config();
-            break;
-        case 5: // PanelTabIndex
-            trapst->panel_tab_idx = value;
-            manufctr->panel_tab_idx = value;
-            update_trap_tab_to_config();
-            break;
-        case 6: // Crate
-            gameadd.object_conf.object_to_door_or_trap[value] = trap_type;
-            gameadd.object_conf.workshop_object_class[value] = TCls_Trap;
-            gameadd.trapdoor_conf.trap_to_object[trap_type] = value;
-            break;
-        case 7: // ManufactureLevel
-            mconf->manufct_level = value;
-            break;
-        case 8: // ManufactureRequired
-            mconf->manufct_required = value;
-            break;
-        case 9: // Shots
-            mconf->shots = value;
-            break;
-        case 10: // TimeBetweenShots
-            mconf->shots_delay = value;
-            break;
-        case 11: // SellingValue
-            mconf->selling_value = value;
-            break;
-        case 12: // Model
-        {
-            struct Objects obj_tmp;
-            gameadd.trap_stats[trap_type].sprite_anim_idx = get_anim_id(context->value->str2, &obj_tmp);
-            refresh_trap_anim(trap_type);
-        }
-            break;
-        case 13: // ModelSize
-            gameadd.trap_stats[trap_type].sprite_size_max = value;
-            refresh_trap_anim(trap_type);
-            break;
-        case 14: // AnimationSpeed
-            gameadd.trap_stats[trap_type].anim_speed = value;
-            refresh_trap_anim(trap_type);
-            break;
-        case 15: // TriggerType
-            gameadd.trap_stats[trap_type].trigger_type = value;
-            break;
-        case 16: // ActivationType
-            gameadd.trap_stats[trap_type].activation_type = value;
-            break;
-        case 17: // EffectType
-            gameadd.trap_stats[trap_type].created_itm_model = value;
-            break;
-        case 18: // Hidden
-            trapst->hidden = value;
-            break;
-        case 19: // TriggerAlarm
-            trapst->notify = value;
-            break;
-        case 20: // Slappable
-            trapst->slappable = value;
-            break;
-        case 21: // Unanimated
-            gameadd.trap_stats[trap_type].unanimated = value;
-            refresh_trap_anim(trap_type);
-            break;
-        default:
-            WARNMSG("Unsupported Trap configuration, variable %d.", context->value->shorts[1]);
-            break;
-    }
-}
-
-static void set_door_configuration_check(const struct ScriptLine* scline)
-{
-    ALLOCATE_SCRIPT_VALUE(scline->command, 0);
-
-    const char *doorname = scline->tp[0];
-    short door_id = get_id(door_desc, doorname);
-    if (door_id == -1)
-    {
-        SCRPTERRLOG("Unknown door, '%s'", doorname);
-        DEALLOCATE_SCRIPT_VALUE
-        return;
-    }
-
-    short doorvar = get_id(door_config_desc, scline->tp[1]);
-    if (doorvar == -1)
-    {
-        SCRPTERRLOG("Unknown door variable");
-        DEALLOCATE_SCRIPT_VALUE
-        return;
-    }
-
-    value->shorts[0] = door_id;
-    value->shorts[1] = doorvar;
-    if (doorvar == 8) // SymbolSprites
-    {
-        char *tmp = malloc(strlen(scline->tp[2]) + strlen(scline->tp[3]) + 3);
-        // Pass two vars along as one merged val like: first\nsecond\m
-        strcpy(tmp, scline->tp[2]);
-        strcat(tmp, "|");
-        strcat(tmp,scline->tp[3]);
-        value->str2 = script_strdup(tmp); // first\0second
-        value->str2[strlen(scline->tp[2])] = 0;
-        free(tmp);
-        if (value->str2 == NULL)
-        {
-            SCRPTERRLOG("Run out script strings space");
-            DEALLOCATE_SCRIPT_VALUE
-            return;
-        }
-    }
-    else if (doorvar != 9) // PointerSprites
-    {
-        if ((scline->np[2] > 0xFFFF) || (scline->np[2] < 0))
-        {
-            SCRPTERRLOG("Value out of range: %d", scline->np[2]);
-            DEALLOCATE_SCRIPT_VALUE
-            return;
-        }
-        value->shorts[2] = (short)scline->np[2];
-    }
-    else
-    {
-        value->str2 = script_strdup(scline->tp[2]);
-        if (value->str2 == NULL)
-        {
-            SCRPTERRLOG("Run out script strings space");
-            DEALLOCATE_SCRIPT_VALUE
-            return;
-        }
-    }
-    SCRIPTDBG(7, "Setting door %s property %s to %d", doorname, doorvar, value->shorts[2]);
-    PROCESS_SCRIPT_VALUE(scline->command);
-}
-
-void set_door_configuration_process(struct ScriptContext *context)
-{
-    long door_type = context->value->shorts[0];
-    struct DoorConfigStats *doorst = get_door_model_stats(door_type);
-    struct ManfctrConfig *mconf = &gameadd.doors_config[door_type];
-    struct ManufactureData *manufctr = get_manufacture_data(gameadd.trapdoor_conf.trap_types_count - 1 + door_type);
-    short value = context->value->shorts[2];
-    switch (context->value->shorts[1])
-    {
-        case 1: // ManufactureLevel
-            mconf->manufct_level = value;
-            break;
-        case 2: // ManufactureRequired
-            mconf->manufct_required = value;
-            break;
-        case 3: // Health
-            if (door_type < DOOR_TYPES_COUNT)
-            {
-                door_stats[door_type][0].health = value;
-                door_stats[door_type][1].health = value;
-            }
-            update_all_door_stats();
-            break;
-        case 4: //SellingValue
-            mconf->selling_value = value;
-            break;
-        case 5: // NametextId
-            doorst->name_stridx = value;
-            break;
-        case 6: // TooltipTextId
-            doorst->tooltip_stridx = value;
-            break;
-        case 7: // Crate
-            gameadd.object_conf.object_to_door_or_trap[value] = door_type;
-            gameadd.object_conf.workshop_object_class[value] = TCls_Door;
-            gameadd.trapdoor_conf.door_to_object[door_type] = value;
-            break;
-        case 8: //SymbolSprites
-            {
-                doorst->bigsym_sprite_idx = get_icon_id(context->value->str2); // First
-                doorst->medsym_sprite_idx = get_icon_id(context->value->str2 + strlen(context->value->str2) + 1); // Second
-                if (doorst->bigsym_sprite_idx < 0)
-                    doorst->bigsym_sprite_idx = bad_icon_id;
-                if (doorst->medsym_sprite_idx < 0)
-                    doorst->medsym_sprite_idx = bad_icon_id;
-                manufctr->bigsym_sprite_idx = doorst->bigsym_sprite_idx;
-                manufctr->medsym_sprite_idx = doorst->medsym_sprite_idx;
-                update_trap_tab_to_config();
-            }
-            break;
-        case 9: // PointerSprites
-            doorst->pointer_sprite_idx = get_icon_id(context->value->str2);
-            if (doorst->pointer_sprite_idx < 0)
-                doorst->pointer_sprite_idx = bad_icon_id;
-            update_trap_tab_to_config();
-            break;
-        case 10: // PanelTabIndex
-            doorst->panel_tab_idx = value;
-            manufctr->panel_tab_idx = value;
-            update_trap_tab_to_config();
-            break;
-        default:
-            WARNMSG("Unsupported Door configuration, variable %d.", context->value->shorts[1]);
-            break;
-    }
-}
 
 
-/**
- * Modifies player's creatures' anger.
- * @param plyr_idx target player
- * @param anger anger value. Use double AnnoyLevel (from creature's config file) to fully piss creature. More for longer calm time
- */
-TbBool script_change_creatures_annoyance(PlayerNumber plyr_idx, ThingModel crmodel, long operation, long anger)
-{
-    SYNCDBG(8, "Starting");
-    struct Dungeon* dungeon = get_players_num_dungeon(plyr_idx);
-    unsigned long k = 0;
-    int i = dungeon->creatr_list_start;
-    while (i != 0)
-    {
-        struct Thing* thing = thing_get(i);
-        TRACE_THING(thing);
-        struct CreatureControl* cctrl = creature_control_get_from_thing(thing);
-        if (thing_is_invalid(thing) || creature_control_invalid(cctrl))
-        {
-            ERRORLOG("Jump to invalid creature detected");
-            break;
-        }
-        i = cctrl->players_next_creature_idx;
-        // Per creature code
-        if (thing->model == crmodel || crmodel == 0)
-        {
-            i = cctrl->players_next_creature_idx;
-            if (operation == SOpr_SET)
-            {
-                anger_set_creature_anger(thing, anger, AngR_Other);
-            }
-            else if (operation == SOpr_INCREASE)
-            {
-                anger_increase_creature_anger(thing, anger, AngR_Other);
-            }
-            else if (operation == SOpr_DECREASE)
-            {
-                anger_reduce_creature_anger(thing, -anger, AngR_Other);
-            }
-            else if (operation == SOpr_MULTIPLY)
-            {
-                anger_set_creature_anger(thing, cctrl->annoyance_level[AngR_Other] * anger, AngR_Other);
-            }
 
-        }
-        // Thing list loop body ends
-        k++;
-        if (k > CREATURES_COUNT)
-        {
-            ERRORLOG("Infinite loop detected when sweeping creatures list");
-            break;
-        }
-    }
-    SYNCDBG(19, "Finished");
-    return true;
-}
 
-static void change_creatures_annoyance_check(const struct ScriptLine* scline)
-{
-    long crtr_id = parse_creature_name(scline->tp[1]);
-    if (crtr_id == CREATURE_NONE)
-    {
-        SCRPTERRLOG("Unknown creature, '%s'", scline->tp[1]);
-        return;
-    }
-    long op_id = get_rid(script_operator_desc, scline->tp[2]);
-    if (op_id == -1)
-    {
-        SCRPTERRLOG("Invalid operation for changing creatures' annoyance: '%s'", scline->tp[2]);
-        return;
-    }
-    command_add_value(Cmd_CHANGE_CREATURES_ANNOYANCE, scline->np[0], crtr_id, op_id, scline->np[3]);
-}
-
-static void change_creatures_annoyance_process(struct ScriptContext* context)
-{
-    for (int i = context->plr_start; i < context->plr_end; i++)
-    {
-        script_change_creatures_annoyance(i, context->value->arg0, context->value->arg1, context->value->arg2);
-    }
-}
 
 static TbBool script_command_param_to_number(char type_chr, struct ScriptLine *scline, int idx, TbBool extended)
 {
@@ -2675,7 +1026,6 @@ short load_script(long lvnum)
 }
 
 
-
 struct Thing *get_creature_in_range_around_any_of_enemy_heart(PlayerNumber plyr_idx, ThingModel crmodel, MapSubtlDelta range)
 {
     int n = GAME_RANDOM(PLAYERS_COUNT);
@@ -2694,7 +1044,6 @@ struct Thing *get_creature_in_range_around_any_of_enemy_heart(PlayerNumber plyr_
     }
     return INVALID_THING;
 }
-
 static struct Thing *script_get_creature_by_criteria(PlayerNumber plyr_idx, long crmodel, long criteria) {
     switch (filter_criteria_type(criteria))
     {
@@ -3128,6 +1477,12 @@ static TbBool is_condition_met(unsigned char cond_idx)
     }
     unsigned long i = gameadd.script.conditions[cond_idx].status;
     return ((i & 0x01) != 0);
+}
+
+static void add_to_party_process(struct ScriptContext *context)
+{
+    struct PartyTrigger* pr_trig = context->pr_trig;
+    add_member_to_party(pr_trig->party_id, pr_trig->creatr_id, pr_trig->crtr_level, pr_trig->carried_gold, pr_trig->objectv, pr_trig->countdown);
 }
 
 static void process_party(struct PartyTrigger* pr_trig)
@@ -4227,76 +2582,6 @@ void script_process_value(unsigned long var_index, unsigned long plr_range_id, l
   }
 }
 
-// TODO: z location
-void find_location_pos(long location, PlayerNumber plyr_idx, struct Coord3d *pos, const char *func_name)
-{
-  struct ActionPoint *apt;
-  struct Thing *thing;
-  unsigned long i = get_map_location_longval(location);
-  memset(pos, 0, sizeof(*pos));
-
-  switch (get_map_location_type(location))
-  {
-    case MLoc_ACTIONPOINT:
-      // Location stores action point index
-      apt = action_point_get(i);
-      if (!action_point_is_invalid(apt))
-      {
-        pos->x.val = apt->mappos.x.val;
-        pos->y.val = apt->mappos.y.val;
-      } else
-        WARNMSG("%s: Action Point %d location not found",func_name,i);
-      break;
-    case MLoc_HEROGATE:
-      thing = find_hero_gate_of_number(i);
-      if (!thing_is_invalid(thing))
-      {
-        *pos = thing->mappos;
-      } else
-        WARNMSG("%s: Hero Gate %d location not found",func_name,i);
-      break;
-    case MLoc_PLAYERSHEART:
-      if (i < PLAYERS_COUNT)
-      {
-        thing = get_player_soul_container(i);
-      } else
-        thing = INVALID_THING;
-      if (!thing_is_invalid(thing))
-      {
-        *pos = thing->mappos;
-      } else
-        WARNMSG("%s: Dungeon Heart location for player %d not found",func_name,i);
-      break;
-    case MLoc_NONE:
-      pos->x.val = 0;
-      pos->y.val = 0;
-      pos->z.val = 0;
-      break;
-    case MLoc_THING:
-      thing = thing_get(i);
-      if (!thing_is_invalid(thing))
-      {
-        *pos = thing->mappos;
-      } else
-        WARNMSG("%s: Thing %d location not found",func_name,i);
-      break;
-    case MLoc_METALOCATION:
-      if (!get_coords_at_meta_action(pos, plyr_idx, i))
-        WARNMSG("%s: Metalocation not found %d",func_name,i);
-      break;
-    case MLoc_CREATUREKIND:
-    case MLoc_OBJECTKIND:
-    case MLoc_ROOMKIND:
-    case MLoc_PLAYERSDUNGEON:
-    case MLoc_APPROPRTDUNGEON:
-    case MLoc_DOORKIND:
-    case MLoc_TRAPKIND:
-    default:
-      WARNMSG("%s: Unsupported location, %lu.",func_name,location);
-      break;
-  }
-  SYNCDBG(15,"From %s; Location %ld, pos(%ld,%ld)",func_name, location, pos->x.stl.num, pos->y.stl.num);
-}
 
 void process_level_script(void)
 {
@@ -4320,191 +2605,6 @@ void process_level_script(void)
 
 
 /******************************************************************************/
-/**
- * Descriptions of script commands for parser.
- * Arguments are: A-string, N-integer, C-creature model, P- player, R- room kind, L- location, O- operator, S- slab kind, X- creature property
- * Lower case letters are optional arguments.
- */
-const struct CommandDesc command_desc[] = {
-  {"CREATE_PARTY",                      "A       ", Cmd_CREATE_PARTY, NULL, NULL},
-  {"ADD_TO_PARTY",                      "ACNNAN  ", Cmd_ADD_TO_PARTY, &add_to_party_check, NULL},
-  {"DELETE_FROM_PARTY",                 "ACN     ", Cmd_DELETE_FROM_PARTY, &delete_from_party_check, NULL},
-  {"ADD_PARTY_TO_LEVEL",                "PAAN    ", Cmd_ADD_PARTY_TO_LEVEL, NULL, NULL},
-  {"ADD_CREATURE_TO_LEVEL",             "PCANNN  ", Cmd_ADD_CREATURE_TO_LEVEL, NULL, NULL},
-  {"ADD_OBJECT_TO_LEVEL",               "AAN     ", Cmd_ADD_OBJECT_TO_LEVEL, NULL, NULL},
-  {"IF",                                "PAON    ", Cmd_IF, NULL, NULL},
-  {"IF_ACTION_POINT",                   "NP      ", Cmd_IF_ACTION_POINT, NULL, NULL},
-  {"ENDIF",                             "        ", Cmd_ENDIF, NULL, NULL},
-  {"SET_HATE",                          "PPN     ", Cmd_SET_HATE, NULL, NULL},
-  {"SET_GENERATE_SPEED",                "N       ", Cmd_SET_GENERATE_SPEED, NULL, NULL},
-  {"REM",                               "        ", Cmd_REM, NULL, NULL},
-  {"START_MONEY",                       "PN      ", Cmd_START_MONEY, NULL, NULL},
-  {"ROOM_AVAILABLE",                    "PRNN    ", Cmd_ROOM_AVAILABLE, NULL, NULL},
-  {"CREATURE_AVAILABLE",                "PCNN    ", Cmd_CREATURE_AVAILABLE, NULL, NULL},
-  {"MAGIC_AVAILABLE",                   "PANN    ", Cmd_MAGIC_AVAILABLE, NULL, NULL},
-  {"TRAP_AVAILABLE",                    "PANN    ", Cmd_TRAP_AVAILABLE, NULL, NULL},
-  {"RESEARCH",                          "PAAN    ", Cmd_RESEARCH, NULL, NULL},
-  {"RESEARCH_ORDER",                    "PAAN    ", Cmd_RESEARCH_ORDER, NULL, NULL},
-  {"COMPUTER_PLAYER",                   "PN      ", Cmd_COMPUTER_PLAYER, NULL, NULL},
-  {"SET_TIMER",                         "PA      ", Cmd_SET_TIMER, NULL, NULL},
-  {"ADD_TUNNELLER_TO_LEVEL",            "PAANNN  ", Cmd_ADD_TUNNELLER_TO_LEVEL, NULL, NULL},
-  {"WIN_GAME",                          "        ", Cmd_WIN_GAME, NULL, NULL},
-  {"LOSE_GAME",                         "        ", Cmd_LOSE_GAME, NULL, NULL},
-  {"SET_FLAG",                          "PAN     ", Cmd_SET_FLAG, NULL, NULL},
-  {"MAX_CREATURES",                     "PN      ", Cmd_MAX_CREATURES, NULL, NULL},
-  {"NEXT_COMMAND_REUSABLE",             "        ", Cmd_NEXT_COMMAND_REUSABLE, NULL, NULL},
-  {"DOOR_AVAILABLE",                    "PANN    ", Cmd_DOOR_AVAILABLE, NULL, NULL},
-  {"DISPLAY_OBJECTIVE",                 "NL      ", Cmd_DISPLAY_OBJECTIVE, &display_objective_check, &display_objective_process},
-  {"DISPLAY_OBJECTIVE_WITH_POS",        "NNN     ", Cmd_DISPLAY_OBJECTIVE_WITH_POS, &display_objective_check, &display_objective_process},
-  {"DISPLAY_INFORMATION",               "NL      ", Cmd_DISPLAY_INFORMATION, NULL, NULL},
-  {"DISPLAY_INFORMATION_WITH_POS",      "NNN     ", Cmd_DISPLAY_INFORMATION_WITH_POS, NULL, NULL},
-  {"ADD_TUNNELLER_PARTY_TO_LEVEL",      "PAAANNN ", Cmd_ADD_TUNNELLER_PARTY_TO_LEVEL, NULL, NULL},
-  {"ADD_CREATURE_TO_POOL",              "CN      ", Cmd_ADD_CREATURE_TO_POOL, NULL, NULL},
-  {"RESET_ACTION_POINT",                "N       ", Cmd_RESET_ACTION_POINT, NULL, NULL},
-  {"SET_CREATURE_MAX_LEVEL",            "PCN     ", Cmd_SET_CREATURE_MAX_LEVEL, NULL, NULL},
-  {"SET_MUSIC",                         "N       ", Cmd_SET_MUSIC, NULL, NULL},
-  {"TUTORIAL_FLASH_BUTTON",             "NN      ", Cmd_TUTORIAL_FLASH_BUTTON, NULL, NULL},
-  {"SET_CREATURE_STRENGTH",             "CN      ", Cmd_SET_CREATURE_STRENGTH, NULL, NULL},
-  {"SET_CREATURE_HEALTH",               "CN      ", Cmd_SET_CREATURE_HEALTH, NULL, NULL},
-  {"SET_CREATURE_ARMOUR",               "CN      ", Cmd_SET_CREATURE_ARMOUR, NULL, NULL},
-  {"SET_CREATURE_FEAR_WOUNDED",         "CN      ", Cmd_SET_CREATURE_FEAR_WOUNDED, NULL, NULL},
-  {"SET_CREATURE_FEAR_STRONGER",        "CN      ", Cmd_SET_CREATURE_FEAR_STRONGER, NULL, NULL},
-  {"SET_CREATURE_FEARSOME_FACTOR",      "CN      ", Cmd_SET_CREATURE_FEARSOME_FACTOR, NULL, NULL},
-  {"SET_CREATURE_PROPERTY",             "CXN     ", Cmd_SET_CREATURE_PROPERTY, NULL, NULL},
-  {"IF_AVAILABLE",                      "PAON    ", Cmd_IF_AVAILABLE, NULL, NULL},
-  {"IF_CONTROLS",                       "PAON    ", Cmd_IF_CONTROLS, NULL, NULL},
-  {"SET_COMPUTER_GLOBALS",              "PNNNNNN ", Cmd_SET_COMPUTER_GLOBALS, NULL, NULL},
-  {"SET_COMPUTER_CHECKS",               "PANNNNN ", Cmd_SET_COMPUTER_CHECKS, NULL, NULL},
-  {"SET_COMPUTER_EVENT",                "PANNNNN ", Cmd_SET_COMPUTER_EVENT, NULL, NULL},
-  {"SET_COMPUTER_PROCESS",              "PANNNNN ", Cmd_SET_COMPUTER_PROCESS, NULL, NULL},
-  {"ALLY_PLAYERS",                      "PPN     ", Cmd_ALLY_PLAYERS, NULL, NULL},
-  {"DEAD_CREATURES_RETURN_TO_POOL",     "N       ", Cmd_DEAD_CREATURES_RETURN_TO_POOL, NULL, NULL},
-  {"BONUS_LEVEL_TIME",                  "Nn      ", Cmd_BONUS_LEVEL_TIME, NULL, NULL},
-  {"QUICK_OBJECTIVE",                   "NAL     ", Cmd_QUICK_OBJECTIVE, NULL, NULL},
-  {"QUICK_INFORMATION",                 "NAL     ", Cmd_QUICK_INFORMATION, NULL, NULL},
-  {"QUICK_OBJECTIVE_WITH_POS",          "NANN    ", Cmd_QUICK_OBJECTIVE_WITH_POS, NULL, NULL},
-  {"QUICK_INFORMATION_WITH_POS",        "NANN    ", Cmd_QUICK_INFORMATION_WITH_POS, NULL, NULL},
-  {"SWAP_CREATURE",                     "AC      ", Cmd_SWAP_CREATURE, NULL, NULL},
-  {"PRINT",                             "A       ", Cmd_PRINT, NULL, NULL},
-  {"MESSAGE",                           "A       ", Cmd_MESSAGE, NULL, NULL},
-  {"PLAY_MESSAGE",                      "PAN     ", Cmd_PLAY_MESSAGE, NULL, NULL},
-  {"ADD_GOLD_TO_PLAYER",                "PN      ", Cmd_ADD_GOLD_TO_PLAYER, NULL, NULL},
-  {"SET_CREATURE_TENDENCIES",           "PAN     ", Cmd_SET_CREATURE_TENDENCIES, NULL, NULL},
-  {"REVEAL_MAP_RECT",                   "PNNNN   ", Cmd_REVEAL_MAP_RECT, NULL, NULL},
-  {"CONCEAL_MAP_RECT",                  "PNNNNa  ", Cmd_CONCEAL_MAP_RECT, &conceal_map_rect_check, &conceal_map_rect_process},
-  {"REVEAL_MAP_LOCATION",               "PNN     ", Cmd_REVEAL_MAP_LOCATION, NULL, NULL},
-  {"LEVEL_VERSION",                     "N       ", Cmd_LEVEL_VERSION, NULL, NULL},
-  {"KILL_CREATURE",                     "PC!AN   ", Cmd_KILL_CREATURE, NULL, NULL},
-  {"COMPUTER_DIG_TO_LOCATION",          "PLL     ", Cmd_COMPUTER_DIG_TO_LOCATION, NULL, NULL},
-  {"USE_POWER_ON_CREATURE",             "PC!APANN", Cmd_USE_POWER_ON_CREATURE, NULL, NULL},
-  {"USE_POWER_AT_POS",                  "PNNANN  ", Cmd_USE_POWER_AT_POS, NULL, NULL},
-  {"USE_POWER_AT_SUBTILE",              "PNNANN  ", Cmd_USE_POWER_AT_POS, NULL, NULL},  //todo: Remove after mapmakers have received time to use USE_POWER_AT_POS
-  {"USE_POWER_AT_LOCATION",             "PNANN   ", Cmd_USE_POWER_AT_LOCATION, NULL, NULL},
-  {"USE_POWER",                         "PAN     ", Cmd_USE_POWER, NULL, NULL},
-  {"USE_SPECIAL_INCREASE_LEVEL",        "PN      ", Cmd_USE_SPECIAL_INCREASE_LEVEL, NULL, NULL},
-  {"USE_SPECIAL_MULTIPLY_CREATURES",    "PN      ", Cmd_USE_SPECIAL_MULTIPLY_CREATURES, NULL, NULL},
-  {"USE_SPECIAL_MAKE_SAFE",             "P       ", Cmd_USE_SPECIAL_MAKE_SAFE, NULL, NULL},
-  {"USE_SPECIAL_LOCATE_HIDDEN_WORLD",   "        ", Cmd_USE_SPECIAL_LOCATE_HIDDEN_WORLD, NULL, NULL},
-  {"CHANGE_CREATURES_ANNOYANCE",        "PC!AN   ", Cmd_CHANGE_CREATURES_ANNOYANCE, &change_creatures_annoyance_check, &change_creatures_annoyance_process},
-  {"ADD_TO_FLAG",                       "PAN     ", Cmd_ADD_TO_FLAG, NULL, NULL},
-  {"SET_CAMPAIGN_FLAG",                 "PAN     ", Cmd_SET_CAMPAIGN_FLAG, NULL, NULL},
-  {"ADD_TO_CAMPAIGN_FLAG",              "PAN     ", Cmd_ADD_TO_CAMPAIGN_FLAG, NULL, NULL},
-  {"EXPORT_VARIABLE",                   "PAA     ", Cmd_EXPORT_VARIABLE, NULL, NULL},
-  {"RUN_AFTER_VICTORY",                 "N       ", Cmd_RUN_AFTER_VICTORY, NULL, NULL},
-  {"LEVEL_UP_CREATURE",                 "PC!AN   ", Cmd_LEVEL_UP_CREATURE, NULL, NULL},
-  {"CHANGE_CREATURE_OWNER",             "PC!AP   ", Cmd_CHANGE_CREATURE_OWNER, NULL, NULL},
-  {"SET_GAME_RULE",                     "AN      ", Cmd_SET_GAME_RULE, NULL, NULL},
-  {"SET_TRAP_CONFIGURATION",            "AANn    ", Cmd_SET_TRAP_CONFIGURATION, &set_trap_configuration_check, &set_trap_configuration_process},
-  {"SET_DOOR_CONFIGURATION",            "AANn    ", Cmd_SET_DOOR_CONFIGURATION, &set_door_configuration_check, &set_door_configuration_process},
-  {"SET_OBJECT_CONFIGURATION",          "AAA     ", Cmd_SET_OBJECT_CONFIGURATION, &set_object_configuration_check, &set_object_configuration_process},
-  {"SET_CREATURE_CONFIGURATION",        "CAAn    ", Cmd_SET_CREATURE_CONFIGURATION, &set_creature_configuration_check, &set_creature_configuration_process},
-  {"SET_SACRIFICE_RECIPE",              "AAA+    ", Cmd_SET_SACRIFICE_RECIPE, &set_sacrifice_recipe_check, &set_sacrifice_recipe_process},
-  {"REMOVE_SACRIFICE_RECIPE",           "A+      ", Cmd_REMOVE_SACRIFICE_RECIPE, &remove_sacrifice_recipe_check, &set_sacrifice_recipe_process},
-  {"SET_BOX_TOOLTIP",                   "NA      ", Cmd_SET_BOX_TOOLTIP, &set_box_tooltip, &null_process},
-  {"SET_BOX_TOOLTIP_ID",                "NN      ", Cmd_SET_BOX_TOOLTIP_ID, &set_box_tooltip_id, &null_process},
-  {"CHANGE_SLAB_OWNER",                 "NNP     ", Cmd_CHANGE_SLAB_OWNER, NULL, NULL},
-  {"CHANGE_SLAB_TYPE",                  "NNS     ", Cmd_CHANGE_SLAB_TYPE, NULL, NULL},
-  {"CREATE_EFFECTS_LINE",               "LLNNNN  ", Cmd_CREATE_EFFECTS_LINE, &create_effects_line_check, &create_effects_line_process},
-  {"IF_SLAB_OWNER",                     "NNP     ", Cmd_IF_SLAB_OWNER, NULL, NULL},
-  {"IF_SLAB_TYPE",                      "NNS     ", Cmd_IF_SLAB_TYPE, NULL, NULL},
-  {"QUICK_MESSAGE",                     "NAA     ", Cmd_QUICK_MESSAGE, NULL, NULL},
-  {"DISPLAY_MESSAGE",                   "NA      ", Cmd_DISPLAY_MESSAGE, NULL, NULL},
-  {"USE_SPELL_ON_CREATURE",             "PC!AAN  ", Cmd_USE_SPELL_ON_CREATURE, NULL, NULL},
-  {"SET_HEART_HEALTH",                  "PN      ", Cmd_SET_HEART_HEALTH, &set_heart_health_check, &set_heart_health_process},
-  {"ADD_HEART_HEALTH",                  "PNn     ", Cmd_ADD_HEART_HEALTH, &add_heart_health_check, &add_heart_health_process},
-  {"CREATURE_ENTRANCE_LEVEL",           "PN      ", Cmd_CREATURE_ENTRANCE_LEVEL, NULL, NULL},
-  {"RANDOMISE_FLAG",                    "PAN     ", Cmd_RANDOMISE_FLAG, NULL, NULL},
-  {"COMPUTE_FLAG",                      "PAAPAN  ", Cmd_COMPUTE_FLAG, NULL, NULL},
-  {"DISPLAY_TIMER",                     "PAn     ", Cmd_DISPLAY_TIMER, &display_timer_check, &display_timer_process},
-  {"ADD_TO_TIMER",                      "PAN     ", Cmd_ADD_TO_TIMER, &add_to_timer_check, &add_to_timer_process},
-  {"ADD_BONUS_TIME",                    "N       ", Cmd_ADD_BONUS_TIME, &add_bonus_time_check, &add_bonus_time_process},
-  {"DISPLAY_VARIABLE",                  "PAnn    ", Cmd_DISPLAY_VARIABLE, &display_variable_check, &display_variable_process},
-  {"DISPLAY_COUNTDOWN",                 "PANn    ", Cmd_DISPLAY_COUNTDOWN, &display_countdown_check, &display_timer_process},
-  {"HIDE_TIMER",                        "        ", Cmd_HIDE_TIMER, &cmd_no_param_check, &hide_timer_process},
-  {"HIDE_VARIABLE",                     "        ", Cmd_HIDE_VARIABLE, &cmd_no_param_check, &hide_variable_process},
-  {"CREATE_EFFECT",                     "AAn     ", Cmd_CREATE_EFFECT, &create_effect_check, &create_effect_process},
-  {"CREATE_EFFECT_AT_POS",              "ANNn    ", Cmd_CREATE_EFFECT_AT_POS, &create_effect_at_pos_check, &create_effect_process},
-  {"HEART_LOST_QUICK_OBJECTIVE",        "NAl     ", Cmd_HEART_LOST_QUICK_OBJECTIVE, &heart_lost_quick_objective_check, &heart_lost_quick_objective_process},
-  {"HEART_LOST_OBJECTIVE",              "Nl      ", Cmd_HEART_LOST_OBJECTIVE, &heart_lost_objective_check, &heart_lost_objective_process},
-  {NULL,                                "        ", Cmd_NONE, NULL, NULL},
-};
-
-const struct CommandDesc dk1_command_desc[] = {
-  {"CREATE_PARTY",                 "A       ", Cmd_CREATE_PARTY, NULL, NULL},
-  {"ADD_TO_PARTY",                 "ACNNAN  ", Cmd_ADD_TO_PARTY, &add_to_party_check, NULL},
-  {"ADD_PARTY_TO_LEVEL",           "PAAN    ", Cmd_ADD_PARTY_TO_LEVEL, NULL, NULL},
-  {"ADD_CREATURE_TO_LEVEL",        "PCANNN  ", Cmd_ADD_CREATURE_TO_LEVEL, NULL, NULL},
-  {"IF",                           "PAON    ", Cmd_IF, NULL, NULL},
-  {"IF_ACTION_POINT",              "NP      ", Cmd_IF_ACTION_POINT, NULL, NULL},
-  {"ENDIF",                        "        ", Cmd_ENDIF, NULL, NULL},
-  {"SET_HATE",                     "PPN     ", Cmd_SET_HATE, NULL, NULL},
-  {"SET_GENERATE_SPEED",           "N       ", Cmd_SET_GENERATE_SPEED, NULL, NULL},
-  {"REM",                          "        ", Cmd_REM, NULL, NULL},
-  {"START_MONEY",                  "PN      ", Cmd_START_MONEY, NULL, NULL},
-  {"ROOM_AVAILABLE",               "PRNN    ", Cmd_ROOM_AVAILABLE, NULL, NULL},
-  {"CREATURE_AVAILABLE",           "PCNN    ", Cmd_CREATURE_AVAILABLE, NULL, NULL},
-  {"MAGIC_AVAILABLE",              "PANN    ", Cmd_MAGIC_AVAILABLE, NULL, NULL},
-  {"TRAP_AVAILABLE",               "PANN    ", Cmd_TRAP_AVAILABLE, NULL, NULL},
-  {"RESEARCH",                     "PAAN    ", Cmd_RESEARCH_ORDER, NULL, NULL},
-  {"COMPUTER_PLAYER",              "PN      ", Cmd_COMPUTER_PLAYER, NULL, NULL},
-  {"SET_TIMER",                    "PA      ", Cmd_SET_TIMER, NULL, NULL},
-  {"ADD_TUNNELLER_TO_LEVEL",       "PAANNN  ", Cmd_ADD_TUNNELLER_TO_LEVEL, NULL, NULL},
-  {"WIN_GAME",                     "        ", Cmd_WIN_GAME, NULL, NULL},
-  {"LOSE_GAME",                    "        ", Cmd_LOSE_GAME, NULL, NULL},
-  {"SET_FLAG",                     "PAN     ", Cmd_SET_FLAG, NULL, NULL},
-  {"MAX_CREATURES",                "PN      ", Cmd_MAX_CREATURES, NULL, NULL},
-  {"NEXT_COMMAND_REUSABLE",        "        ", Cmd_NEXT_COMMAND_REUSABLE, NULL, NULL},
-  {"DOOR_AVAILABLE",               "PANN    ", Cmd_DOOR_AVAILABLE, NULL, NULL},
-  {"DISPLAY_OBJECTIVE",            "NA      ", Cmd_DISPLAY_OBJECTIVE, &display_objective_check, &display_objective_process},
-  {"DISPLAY_OBJECTIVE_WITH_POS",   "NNN     ", Cmd_DISPLAY_OBJECTIVE_WITH_POS, &display_objective_check, &display_objective_process},
-  {"DISPLAY_INFORMATION",          "N       ", Cmd_DISPLAY_INFORMATION, NULL, NULL},
-  {"DISPLAY_INFORMATION_WITH_POS", "NNN     ", Cmd_DISPLAY_INFORMATION_WITH_POS, NULL, NULL},
-  {"ADD_TUNNELLER_PARTY_TO_LEVEL", "PAAANNN ", Cmd_ADD_TUNNELLER_PARTY_TO_LEVEL, NULL, NULL},
-  {"ADD_CREATURE_TO_POOL",         "CN      ", Cmd_ADD_CREATURE_TO_POOL, NULL, NULL},
-  {"RESET_ACTION_POINT",           "N       ", Cmd_RESET_ACTION_POINT, NULL, NULL},
-  {"SET_CREATURE_MAX_LEVEL",       "PCN     ", Cmd_SET_CREATURE_MAX_LEVEL, NULL, NULL},
-  {"SET_MUSIC",                    "N       ", Cmd_SET_MUSIC, NULL, NULL},
-  {"TUTORIAL_FLASH_BUTTON",        "NN      ", Cmd_TUTORIAL_FLASH_BUTTON, NULL, NULL},
-  {"SET_CREATURE_STRENGTH",        "CN      ", Cmd_SET_CREATURE_STRENGTH, NULL, NULL},
-  {"SET_CREATURE_HEALTH",          "CN      ", Cmd_SET_CREATURE_HEALTH, NULL, NULL},
-  {"SET_CREATURE_ARMOUR",          "CN      ", Cmd_SET_CREATURE_ARMOUR, NULL, NULL},
-  {"SET_CREATURE_FEAR",            "CN      ", Cmd_SET_CREATURE_FEAR_WOUNDED, NULL, NULL},
-  {"IF_AVAILABLE",                 "PAON    ", Cmd_IF_AVAILABLE, NULL, NULL},
-  {"SET_COMPUTER_GLOBALS",         "PNNNNNN ", Cmd_SET_COMPUTER_GLOBALS, NULL, NULL},
-  {"SET_COMPUTER_CHECKS",          "PANNNNN ", Cmd_SET_COMPUTER_CHECKS, NULL, NULL},
-  {"SET_COMPUTER_EVENT",           "PANN    ", Cmd_SET_COMPUTER_EVENT, NULL, NULL},
-  {"SET_COMPUTER_PROCESS",         "PANNNNN ", Cmd_SET_COMPUTER_PROCESS, NULL, NULL},
-  {"ALLY_PLAYERS",                 "PP      ", Cmd_ALLY_PLAYERS, NULL, NULL},
-  {"DEAD_CREATURES_RETURN_TO_POOL","N       ", Cmd_DEAD_CREATURES_RETURN_TO_POOL, NULL, NULL},
-  {"BONUS_LEVEL_TIME",             "N       ", Cmd_BONUS_LEVEL_TIME, NULL, NULL},
-  {"QUICK_OBJECTIVE",              "NAA     ", Cmd_QUICK_OBJECTIVE, NULL, NULL},
-  {"QUICK_INFORMATION",            "NA      ", Cmd_QUICK_INFORMATION, NULL, NULL},
-  {"SWAP_CREATURE",                "AC      ", Cmd_SWAP_CREATURE, NULL, NULL},
-  {"PRINT",                        "A       ", Cmd_PRINT, NULL, NULL},
-  {"MESSAGE",                      "A       ", Cmd_MESSAGE, NULL, NULL},
-  {"LEVEL_VERSION",                "N       ", Cmd_LEVEL_VERSION, NULL, NULL},
-  {NULL,                           "        ", Cmd_NONE, NULL, NULL},
-};
 
 /******************************************************************************/
 #ifdef __cplusplus
