+++ src/lvl_script_lib.h
@@ -174,6 +174,82 @@ struct CommandDesc { // sizeof = 14 // originally was 13
   void (*process_fn)(struct ScriptContext *context); // called from value or from
 };
 
+enum ScriptVariables {
+  SVar_MONEY                           =  1,
+  SVar_GAME_TURN                       =  5,
+  SVar_BREAK_IN                        =  6,
+  SVar_CREATURE_NUM                    =  7,
+  SVar_TOTAL_DIGGERS                   =  8,
+  SVar_TOTAL_CREATURES                 =  9,
+  SVar_TOTAL_RESEARCH                  = 10,
+  SVar_TOTAL_DOORS                     = 11,
+  SVar_TOTAL_AREA                      = 12,
+  SVar_TOTAL_CREATURES_LEFT            = 13,
+  SVar_CREATURES_ANNOYED               = 14,
+  SVar_BATTLES_LOST                    = 15,
+  SVar_BATTLES_WON                     = 16,
+  SVar_ROOMS_DESTROYED                 = 17,
+  SVar_SPELLS_STOLEN                   = 18,
+  SVar_ACTION_POINT_TRIGGERED          = 19,
+  SVar_GOLD_POTS_STOLEN                = 20,
+  SVar_TIMER                           = 21,
+  SVar_DUNGEON_DESTROYED               = 22,
+  SVar_TIMES_BROKEN_INTO               = 23,
+  SVar_TOTAL_GOLD_MINED                = 24,
+  SVar_FLAG                            = 25,
+  SVar_ROOM_SLABS                      = 26,
+  SVar_DOORS_DESTROYED                 = 27,
+  SVar_CREATURES_SCAVENGED_LOST        = 28,
+  SVar_CREATURES_SCAVENGED_GAINED      = 29,
+  SVar_AVAILABLE_MAGIC                 = 30,
+  SVar_AVAILABLE_TRAP                  = 31,
+  SVar_AVAILABLE_DOOR                  = 32,
+  SVar_AVAILABLE_ROOM                  = 33,
+  SVar_AVAILABLE_CREATURE              = 34,
+  SVar_CONTROLS_CREATURE               = 35,
+  SVar_CONTROLS_TOTAL_CREATURES        = 36,
+  SVar_CONTROLS_TOTAL_DIGGERS          = 37,
+  SVar_ALL_DUNGEONS_DESTROYED          = 38,
+  SVar_DOOR_NUM                        = 39,
+  SVar_TRAP_NUM                        = 40,
+  SVar_GOOD_CREATURES                  = 41,
+  SVar_EVIL_CREATURES                  = 42,
+  SVar_CONTROLS_GOOD_CREATURES         = 43,
+  SVar_CONTROLS_EVIL_CREATURES         = 44,
+  SVar_CAMPAIGN_FLAG                   = 45,
+  SVar_SLAB_OWNER                      = 46,
+  SVar_SLAB_TYPE                       = 47,
+  SVar_HEART_HEALTH                    = 48,
+  SVar_GHOSTS_RAISED                   = 49,
+  SVar_SKELETONS_RAISED                = 50,
+  SVar_VAMPIRES_RAISED                 = 51,
+  SVar_CREATURES_CONVERTED             = 52,
+  SVar_TIMES_ANNOYED_CREATURE          = 53,
+  SVar_TIMES_TORTURED_CREATURE         = 54,
+  SVar_TOTAL_DOORS_MANUFACTURED        = 55,
+  SVar_TOTAL_TRAPS_MANUFACTURED        = 56,
+  SVar_TOTAL_MANUFACTURED              = 57,
+  SVar_TOTAL_TRAPS_USED                = 58,
+  SVar_TOTAL_DOORS_USED                = 59,
+  SVar_KEEPERS_DESTROYED               = 60,
+  SVar_CREATURES_SACRIFICED            = 61, // Total
+  SVar_CREATURES_FROM_SACRIFICE        = 62, // Total
+  SVar_TIMES_LEVELUP_CREATURE          = 63,
+  SVar_TOTAL_SALARY                    = 64,
+  SVar_CURRENT_SALARY                  = 65,
+  SVar_BOX_ACTIVATED                   = 66,
+  SVar_SACRIFICED                      = 67,  // Per model
+  SVar_REWARDED                        = 68,  // Per model
+  SVar_EVIL_CREATURES_CONVERTED        = 69,
+  SVar_GOOD_CREATURES_CONVERTED        = 70,
+  SVar_TRAPS_SOLD                      = 71,
+  SVar_DOORS_SOLD                      = 72,
+  SVar_MANUFACTURED_SOLD               = 73,
+  SVar_MANUFACTURE_GOLD                = 74,
+  SVar_TOTAL_SCORE                     = 75,
+  SVar_BONUS_TIME                      = 76,
+ };
+
 extern const struct CommandDesc command_desc[];
 
 extern const struct CommandDesc subfunction_desc[];
@@ -194,6 +270,11 @@ extern const struct NamedCommand trap_config_desc[];
 extern const struct NamedCommand gui_button_group_desc[];
 extern const struct NamedCommand campaign_flag_desc[];
 extern const struct NamedCommand script_operator_desc[];
+
+
+
+// 1/4 turn minimal
+#define FX_LINE_TIME_PARTS 4
 
 
 
@@ -233,6 +314,21 @@ char get_player_number_from_value(const char* txt);
         } \
     } \
     command_init_value(value, var_index, plr_range_id);
+
+#define DEALLOCATE_SCRIPT_VALUE \
+    if (value != &tmp_value) \
+    {                           \
+        value->flags = TrgF_DISABLED; \
+        gameadd.script.values_num--; \
+    }
+
+    void script_process_value(unsigned long var_index, unsigned long plr_range_id, long val2, long val3, long val4, struct ScriptValue *value);
+
+#define PROCESS_SCRIPT_VALUE(cmd) \
+    if ((get_script_current_condition() == CONDITION_ALWAYS) && (next_command_reusable == 0)) \
+    { \
+        script_process_value(cmd, 0, 0, 0, 0, value); \
+    }
 
 
 #ifdef __cplusplus
